Add files via upload
I'm uploading what I've got right now, it still needs to get comments and we'll have to work to merge it into the rest of the project, I'm just uploading this now so that you can work on my part if you need to.
The CSVParser isn't going to be part of the final project most likely, it's just a tool to convert scpscrape.csv (the output of the web scraper I used) into usable .txt files. Make sure there's an existing /data/ folder and run it and it'll turn scpscrape.csv into 1000 individual text documents.
I'll be on the road for about 5 hours but when I arrive I'll get to work on the project again!